requirements.txt
# Web scraping and HTTP
requests>=2.31.0
beautifulsoup4>=4.12.0
feedparser>=6.0.0

# Google Gemini AI
google-generativeai>=0.3.0

# Configuration
pyyaml>=6.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0

# Email formatting
jinja2>=3.1.0

# Web interface
flask>=3.0.0
gunicorn>=21.0.0

# Utilities
python-dateutil>=2.8.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0
